+++ src/ArduinoJson/Object/MemberProxy.hpp
@@ -39,7 +39,7 @@ class MemberProxy
     return *this;
   }
 
-  template <typename T, typename = enable_if_t<!is_const<T>::value>>
+  template <typename T, enable_if_t<!is_const<T>::value, int> = 0>
   MemberProxy& operator=(T* src) {
     this->set(src);
     return *this;